Implemented URLManagerException class.
Most likely not a complete yet, but I feel like its what I would like it to be.

I am still probably going to be looking at other methods of doing this, but only if they are in pure JS and not using any other frameworks.

Next time I will go over URLParameter a little more and comment it.
It should be fine as is right now, but just to be sure. It still hasno comments, which is something I plan on adding.

URLManager main class has not been started yet. The plan is to do so in the next session.