Full implementation complete.
Although the whole thing should function as is, it is not tested and not all exceptions are complete either.

I have completed the URLParameter class but not tested it.

I have mostly completed the URLManager class, but not tested it.
There may be some parts that still need to be updated.

Do not use yet.